Add varint unit tests.
Make line_num i32 (2,147,483,647 is more than enough).

+++ parse/migrations/2024-05-19-114528_create/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE lines(
     file_id    INTEGER   NOT NULL,
-    line_num   BIGINT    NOT NULL,
+    line_num   INTEGER   NOT NULL,
     -- Log Level (Info, Verbose, etc.)
     level      INTEGER   NOT NULL,
     timestamp  TIMESTAMP NOT NULL,
